Scrolling is enough for consent, removed user agent sniffing, click any link to consent now default true.
Removed all the jquery rules for mobile, included user agent sniffing
that isn't reliable. Added a scrollToConsent setting, because it's
enough for having consent in installing cookies. I wanted also to create
a "Click any link to consent" function, but I notice it's already there,
so I put it on "true" by default.